+++ src/Application/app.scss
@@ -1,5 +1,5 @@
 @import 'Components/NewsRoom/NewsRoom';
-@import 'Components/Store/StoreItem';
+@import 'Components/Shop/ShopItem';
 
 .side-panel {
   padding-top: 3em;
